Fixed audio issues between chunks
STill aproblem:
Chunks will skip towards beginning of episode. Generally after first set of commercials?

Possibly a problem still but not sure: Ads not playing at the beginning of a new stream (if the ads are too close to the beginning???)
Volume goes up and down for each chunk (Different volume per chunk) ??? Hard to tell, it might not be happening anymore.